# src/main/java/com/bistro/infrastructure/web/dto/request/OrderItemRequest.java
package com.bistro.infrastructure.web.dto.request;

import jakarta.validation.constraints.Min;
import jakarta.validation.constraints.NotNull;

public record OrderItemRequest(
    @NotNull Long productId,
    @Min(1) int quantity
) {}