# src/main/java/com/bistro/infrastructure/web/dto/request/OrderRequest.java
package com.bistro.infrastructure.web.dto.request;

import jakarta.validation.Valid;
import jakarta.validation.constraints.NotEmpty;
import jakarta.validation.constraints.NotNull;
import java.util.List;

public record OrderRequest(
    @NotNull Integer tableNumber,
    @NotEmpty @Valid List<OrderItemRequest> items
) {}