# src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  title: Bistro API
  version: 1.0.0

tags:
  - name: Orders
    description: Place and retrieve orders
  - name: Products
    description: Browse the product catalogue

paths:
  /orders:
    post:
      tags: [Orders]
      summary: Place a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
            example:
              tableNumber: 5
              items:
                - productId: 1
                  quantity: 2
                - productId: 3
                  quantity: 1
      responses:
        '201':
          description: Order created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          description: Malformed JSON
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: Malformed JSON request
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: 'Product not found: 99'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: 'items: must not be empty'

  /orders/{id}:
    get:
      tags: [Orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: 'Order not found: 99'

  /products:
    get:
      tags: [Products]
      summary: List all products
      responses:
        '200':
          description: Product list returned (may be empty)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'

  /products/{id}:
    get:
      tags: [Products]
      summary: Get a product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          example: 1
      responses:
        '200':
          description: Product found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
        '404':
          description: Product not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                message: 'Product not found: 99'

components:
  schemas:
    OrderRequest:
      type: object
      required: [tableNumber, items]
      properties:
        tableNumber:
          type: integer
          example: 5
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemRequest'

    OrderItemRequest:
      type: object
      required: [productId, quantity]
      properties:
        productId:
          type: integer
          format: int64
          example: 1
        quantity:
          type: integer
          minimum: 1
          example: 2

    OrderResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        tableNumber:
          type: integer
          example: 5
        orderTime:
          type: string
          format: date-time
          example: '2026-05-10T12:30:00'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        subtotal:
          type: number
          example: 24.50
        discounts:
          type: array
          items:
            $ref: '#/components/schemas/OrderDiscount'
        total:
          type: number
          example: 22.05
        receipt:
          type: string
          example: "Table 5 | 2x Margherita Pizza @ 12.50 | Subtotal: 25.00 | Happy Hour -2.50 | Total: 22.50"

    OrderItemResponse:
      type: object
      properties:
        productName:
          type: string
          example: Margherita Pizza
        quantity:
          type: integer
          example: 2
        unitPrice:
          type: number
          example: 12.50

    OrderDiscount:
      type: object
      properties:
        description:
          type: string
          example: Happy Hour
        amount:
          type: number
          example: 2.50

    Product:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Margherita Pizza
        price:
          type: number
          example: 12.50
        category:
          type: string
          enum: [PIZZA, DRINK, DESSERT, SNACK]
          example: PIZZA

    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          example: 'Product not found: 99'